メーター生成を meter.py に分離
